+++ nmtpytorch/datasets/__init__.py
@@ -5,3 +5,4 @@ from .text import TextDataset
 from .multiparallel import MultiParallelDataset
 from .onehot import OneHotDataset
 from .multi30kraw import Multi30kRawDataset
+from .numpy import NumpyDataset